companion: Cleanup and split into one file per command
