+++ doc/release/master/cleanup_companion.md
@@ -1,0 +1,11 @@
+companion_env {#master}
+-------------
+
+### Tools
+
+#### `yarp`
+
+* The `regression` subcommand was removed.
+* The `--help`, `-h` and `--version` arguments are now accepted.
+* The `qos` subcommand is accepted as an alias to `priority-qos`
+* The `sched` subcommand is accepted as an alias to `priority-sched`
