+++ src/libYARP_companion/src/yarp/companion/impl/Companion.cmdEnv.cpp
